+++ libraries/typescript/packages/mcp-use/CHANGELOG.md
@@ -1,5 +1,13 @@
 # mcp-use
 
+## 1.21.2-canary.0
+
+### Patch Changes
+
+- Updated dependencies [76320f0]
+  - @mcp-use/inspector@0.24.2-canary.0
+  - @mcp-use/cli@2.18.1-canary.0
+
 ## 1.21.1
 
 ### Patch Changes